% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_ops.R
\name{register_dubious_syntaxes}
\alias{register_dubious_syntaxes}
\title{Register Dubious Syntaxes}
\usage{
register_dubious_syntaxes(syntaxes)
}
\arguments{
\item{syntaxes}{a character vector of the syntaxes to support}
}
\description{
To use a dubious syntax in a package, use this function in the definition of \code{.onAttach}
}
\examples{
\dontrun{
# define your syntax as you would define a normal function
`?add> {x} : {y}` <- function(x, y) x + y

# register the syntax in your .onAttach definition
.onAttach <- function(libname, pkgname) {
doubt::register_dubious_syntaxes("?add> {x} : {y}")
invisible()
}
}
}
